# docs/swagger.yaml
basePath: /
definitions:
...
  abude-backend_internal_pkg_inventories_category.Category:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      parent:
        $ref: '#/definitions/abude-backend_internal_pkg_inventories_category.Category'
      updatedAt:
        type: string
    type: object
...
  common.BasicResponse:
    properties:
      message:
        type: string
    type: object
...
  common.GeneralResponse:
    properties:
      message:
        type: string
      result: {}
    type: object
  common.PaginatedResponse:
    properties:
      metadata:
        $ref: '#/definitions/pagination.Metadata'
      result:
        items: {}
        type: array
    type: object
  company.Company:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      region:
        type: string
      updatedAt:
        type: string
    type: object
...
  pagination.Metadata:
    properties:
      count:
        type: integer
      hasNext:
        type: boolean
      hasPrev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  product.Ingredient:
    properties:
      base:
        $ref: '#/definitions/product.Product'
      createdAt:
        type: string
      id:
        type: integer
      ingredient:
        $ref: '#/definitions/product.Product'
      quantity:
        type: number
      updatedAt:
        type: string
    type: object
...
  product.Product:
    properties:
      category:
        $ref: '#/definitions/abude-backend_internal_pkg_inventories_category.Category'
      company:
        $ref: '#/definitions/company.Company'
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      ingredients:
        items:
          $ref: '#/definitions/product.Ingredient'
        type: array
      isDefault:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock:
        type: boolean
      type:
        enum:
        - purchase
        - sale
        type: string
      unit:
        type: string
      updatedAt:
        type: string
    type: object
...
  purchase.Purchase:
    properties:
      code:
        type: string
      createdAt:
        type: string
      date:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/purchase.PurchaseItem'
        type: array
      note:
        type: string
      status:
        enum:
        - approved
        - accepted
        - canceled
        type: string
      supplier:
        $ref: '#/definitions/supplier.Supplier'
      total:
        type: integer
      type:
        enum:
        - debit
        - credit
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  purchase.PurchaseDTO:
    properties:
      date:
        format: date-time
        type: string
      items:
        items:
          $ref: '#/definitions/purchase.PurchaseItemDTO'
        type: array
      note:
        type: string
      source:
        enum:
        - outlet
        - warehouse
        type: string
      sourceId:
        type: integer
      supplier:
        type: integer
      type:
        enum:
        - debit
        - credit
        type: string
    required:
    - items
    - source
    - sourceId
    - type
    type: object
  purchase.PurchaseItem:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      price:
        type: integer
      product:
        $ref: '#/definitions/product.Product'
      quantity:
        type: integer
      status:
        type: boolean
      total:
        type: integer
      updatedAt:
        type: string
    type: object
  purchase.PurchaseItemDTO:
    properties:
      price:
        type: integer
      product:
        type: integer
      quantity:
        type: integer
    required:
    - product
    - quantity
    type: object
  purchase.PurchaseSummary:
    properties:
      date:
        type: string
      id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      total:
        type: number
    type: object
  sale.Sale:
    properties:
      code:
        type: string
      createdAt:
        type: string
      customer:
        type: string
      date:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/sale.SaleItem'
        type: array
      note:
        type: string
      status:
        enum:
        - approved
        - accepted
        - canceled
        type: string
      total:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  sale.SaleDTO:
    properties:
      customer:
        type: string
      date:
        format: date-time
        type: string
      items:
        items:
          $ref: '#/definitions/sale.SaleItemDTO'
        type: array
      note:
        type: string
      source:
        enum:
        - outlet
        - warehouse
        type: string
      sourceId:
        type: integer
      status:
        enum:
        - accepted
        - canceled
        - approved
        type: string
    required:
    - customer
    - items
    - source
    - sourceId
    type: object
  sale.SaleItem:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      price:
        type: integer
      product:
        $ref: '#/definitions/product.Product'
      quantity:
        type: integer
      status:
        type: boolean
      total:
        type: integer
      updatedAt:
        type: string
    type: object
  sale.SaleItemDTO:
    properties:
      price:
        type: integer
      product:
        type: integer
      quantity:
        type: integer
    required:
    - product
    - quantity
    type: object
  sale.SaleSummary:
    properties:
      date:
        type: string
      id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      total:
        type: number
    type: object
...
  supplier.Supplier:
    properties:
      company:
        $ref: '#/definitions/company.Company'
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  user.User:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        enum:
        - superadmin
        - owner
        - employee
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
info:
  contact:
    email: infamous0192@gmail.com
    name: Dwa Meizadewa
  description: Aplikasi Backend (API) untuk project Abude
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Abude Backend
  version: 2.0.0
paths:
...
  /api/purchase:
    get:
      consumes:
      - application/json
      parameters:
      - format: date-time
        in: query
        name: endDate
        type: string
      - in: query
        name: limit
        type: integer
      - description: Outlet ID
        in: query
        name: outlet
        type: integer
      - in: query
        name: page
        type: integer
      - format: date-time
        in: query
        name: startDate
        type: string
      - collectionFormat: csv
        in: query
        items:
          enum:
          - accepted
          - approved
          - canceled
          type: string
        name: status
        type: array
      - description: User ID
        in: query
        name: user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/purchase.Purchase'
                  type: array
              type: object
      security:
      - JWT: []
      summary: Get All Purchase
      tags:
      - Purchases
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/purchase.PurchaseDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/purchase.Purchase'
              type: object
      security:
      - JWT: []
      summary: Create Purchase
      tags:
      - Purchases
  /api/purchase/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Purchase ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/purchase.Purchase'
              type: object
      security:
      - JWT: []
      summary: Delete Purchase
      tags:
      - Purchases
    get:
      consumes:
      - application/json
      parameters:
      - description: Purchase ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/purchase.Purchase'
      security:
      - JWT: []
      summary: Get One Purchase
      tags:
      - Purchases
    put:
      consumes:
      - application/json
      parameters:
      - description: Purchase ID
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/purchase.PurchaseDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/purchase.Purchase'
              type: object
      security:
      - JWT: []
      summary: Update Purchase
      tags:
      - Purchases
  /api/purchase/{id}/cancel:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Purchase ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BasicResponse'
      security:
      - JWT: []
      summary: Cancel Purchase
      tags:
      - Purchases
  /api/purchase/summary:
    get:
      consumes:
      - application/json
      parameters:
      - format: date-time
        in: query
        name: endDate
        type: string
      - description: Outlet ID
        in: query
        name: outlet
        type: integer
      - format: date-time
        in: query
        name: startDate
        type: string
      - collectionFormat: csv
        in: query
        items:
          enum:
          - accepted
          - approved
          - canceled
          type: string
        name: status
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/purchase.PurchaseSummary'
            type: array
      security:
      - JWT: []
      summary: Get Purchases Summary
      tags:
      - Purchases
  /api/sale:
    get:
      consumes:
      - application/json
      parameters:
      - format: date-time
        in: query
        name: endDate
        type: string
      - in: query
        name: limit
        type: integer
      - description: Outlet ID
        in: query
        name: outlet
        type: integer
      - in: query
        name: page
        type: integer
      - format: date-time
        in: query
        name: startDate
        type: string
      - collectionFormat: csv
        in: query
        items:
          enum:
          - accepted
          - approved
          - canceled
          type: string
        name: status
        type: array
      - description: User ID
        in: query
        name: user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/sale.Sale'
                  type: array
              type: object
      security:
      - JWT: []
      summary: Get All Sale
      tags:
      - Sales
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sale.SaleDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/sale.Sale'
              type: object
      security:
      - JWT: []
      summary: Create Sale
      tags:
      - Sales
  /api/sale/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Sale ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/sale.Sale'
              type: object
      security:
      - JWT: []
      summary: Delete Sale
      tags:
      - Sales
    get:
      consumes:
      - application/json
      parameters:
      - description: Sale ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sale.Sale'
      security:
      - JWT: []
      summary: Get One Sale
      tags:
      - Sales
    put:
      consumes:
      - application/json
      parameters:
      - description: Sale ID
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sale.SaleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/sale.Sale'
              type: object
      security:
      - JWT: []
      summary: Update Sale
      tags:
      - Sales
  /api/sale/{id}/cancel:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Sale ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BasicResponse'
      security:
      - JWT: []
      summary: Cancel Sale
      tags:
      - Sales
  /api/sale/summary:
    get:
      consumes:
      - application/json
      parameters:
      - format: date-time
        in: query
        name: endDate
        type: string
      - description: Outlet ID
        in: query
        name: outlet
        type: integer
      - format: date-time
        in: query
        name: startDate
        type: string
      - collectionFormat: csv
        in: query
        items:
          enum:
          - accepted
          - approved
          - canceled
          type: string
        name: status
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sale.SaleSummary'
            type: array
      security:
      - JWT: []
      summary: Get Sales Summary
      tags:
      - Sales
...
schemes:
- https
- http
securityDefinitions:
  JWT:
    description: Authorization for JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"